% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_field_names.R
\name{add_field_names}
\alias{add_field_names}
\title{Add names within HL7 segments - field names}
\usage{
add_field_names(hl7_list_element, dev_msg = FALSE)
}
\arguments{
\item{hl7_list_element}{a list, that from one after HL7 parsing has been conducted}

\item{dev_msg}{boolean, development message - reveal field which still require names}
}
\description{
This function examines the lengths of elements within a parsed HL7 and if the
length is > 1, it will compare the element name to a list of gathered names in
\code{segment_names}. If names are found, the field names will be applied
}
