% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse_hl7.R
\name{browse_hl7}
\alias{browse_hl7}
\title{Browse a parsed HL7}
\usage{
browse_hl7(x)
}
\arguments{
\item{x}{a HL7 file to parse}
}
\description{
Browse a parsed HL7 in the viewer pane. This function uses the details and summary
HTML tags for effect
}
