% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_hl7.R
\name{build_hl7}
\alias{build_hl7}
\title{Build HL7 character vector from segment functions}
\usage{
build_hl7(...)
}
\arguments{
\item{...}{segment functions MSH(), OBR(), etc.}
}
\value{
character vector of length equal to the number of segment functions and
of class 'hl7vector'
}
\description{
Build a character vector from segment functions for export as HL7 text files
The function only accepts objects from the segment functions MSH(), OBR(), etc.
The first segment must be MSH().
}
\examples{
build_hl7(MSH(), PID())

}
