% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime_to_hl7_datetime.R
\name{datetime_to_hl7_datetime}
\alias{datetime_to_hl7_datetime}
\title{Convert a date or datetime to HL7 datetime}
\usage{
datetime_to_hl7_datetime(x)
}
\arguments{
\item{x}{date/datetime, object with class in c("Date", "POSIXct", "POSIXt")}
}
\description{
Simple wrapper to convert to date/datetime without any separators
}
