% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_field_separators.R
\name{extract_field_separators}
\alias{extract_field_separators}
\title{Extract field separators from with the MSH segment of a HL7 message}
\usage{
extract_field_separators(hl7_char)
}
\arguments{
\item{hl7_char}{a character vector, the lines from a HL7 message}
}
\description{
These are usually default, but if not, they'l; be captured and used for parsing
a message
}
