% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FHS.R
\name{FHS}
\alias{FHS}
\title{Segment FHS - File Header Segment}
\usage{
FHS(
  FHS.1_FileFieldSeparator = "|",
  FHS.2_FileEncodingCharacters = "",
  FHS.3_FileSendingApplication = "",
  FHS.4_FileSendingFacility = "",
  FHS.5_FileReceivingApplication = "",
  FHS.6_FileReceivingFacility = "",
  FHS.7_FileCreationDateTime = "",
  FHS.8_FileSecurity = "",
  FHS.9_FileNameID = "",
  FHS.10_FileHeaderComment = "",
  FHS.11_FileControlID = "",
  FHS.12_ReferenceFileControlID = "",
  .version = c("2.3.1", "2.4"),
  .trim = TRUE
)
}
\arguments{
\item{FHS.1_FileFieldSeparator}{= "|",}

\item{FHS.2_FileEncodingCharacters}{= "",}

\item{FHS.3_FileSendingApplication}{= "",}

\item{FHS.4_FileSendingFacility}{= "",}

\item{FHS.5_FileReceivingApplication}{= "",}

\item{FHS.6_FileReceivingFacility}{= "",}

\item{FHS.7_FileCreationDateTime}{= "",}

\item{FHS.8_FileSecurity}{= "",}

\item{FHS.9_FileNameID}{= "",}

\item{FHS.10_FileHeaderComment}{= "",}

\item{FHS.11_FileControlID}{= "",}

\item{FHS.12_ReferenceFileControlID}{= "",}

\item{.version}{= c("2.3.1", "2.4"),}

\item{.trim}{= TRUE}
}
\value{
character vector
}
\description{
Convert input to a FHS segment for HL7 2.3.1 or 2.4. Outputs a character vector of
length 1 with input values pipe delimited.
}
