% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hl7_version.R
\name{get_hl7_version}
\alias{get_hl7_version}
\title{Extract HL7 version ID from MSH segment}
\usage{
get_hl7_version(hl7_char)
}
\arguments{
\item{hl7_char}{character vector, character vector of segments. Only the first line is interrogated.}
}
\description{
Extract element 12.1 from the MSH segment. Version ID > Version ID
}
