% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSH.R
\name{MSH}
\alias{MSH}
\alias{EncodingCharactersComponents}
\alias{SendingFacilityComponents}
\alias{MessageTypeComponents}
\alias{VersionIDComponents}
\title{Segment MSH: Message Header}
\usage{
MSH(
  EncodingCharacters = EncodingCharactersComponents(),
  SendingApplication = "",
  SendingFacility = SendingFacilityComponents(),
  ReceivingApplication = "",
  ReceivingFacility = "",
  DateTimeOfMessage = "",
  Security = "",
  MessageType = MessageTypeComponents(),
  MessageControlID = "",
  ProcessingID = "",
  VersionID = "",
  SequenceNumber = "",
  ContinuationPointer = "",
  AcceptAcknowledgmentType = "",
  ApplicationAcknowledgmentType = "",
  CountryCode = "",
  CharacterSet = "",
  PrincipalLanguageOfMessage = "",
  AlternateCharacterSetHandlingScheme = "",
  ConformanceStatementID = "",
  .FieldSeparator = "|",
  .version = c("2.3.1", "2.4"),
  .trim = TRUE
)

EncodingCharactersComponents(
  componentSeparator = "^",
  repetitionSeparator = "~",
  escapeCharacter = "\\\\",
  subcomponentSeparator = "&"
)

SendingFacilityComponents(
  namespace_ID = "",
  universalID = "",
  universalIDtype = "",
  .sep = "^",
  .trim = TRUE
)

MessageTypeComponents(
  messageType = "",
  triggerEvent = "",
  messageStructure = "",
  .sep = "^",
  .trim = TRUE
)

VersionIDComponents(
  versionId = "",
  internationalizationCode = "",
  internationalversionId = "",
  .sep = "^",
  .trim = TRUE
)
}
\arguments{
\item{EncodingCharacters}{= EncodingCharactersComponents(),}

\item{SendingApplication}{= "",}

\item{SendingFacility}{= SendingFacilityComponents(),}

\item{ReceivingApplication}{= "",}

\item{ReceivingFacility}{= "",}

\item{DateTimeOfMessage}{= "",}

\item{Security}{= "",}

\item{MessageType}{= MessageTypeComponents(),}

\item{MessageControlID}{= "",}

\item{ProcessingID}{= "",}

\item{VersionID}{= "",}

\item{SequenceNumber}{= "",}

\item{ContinuationPointer}{= "",}

\item{AcceptAcknowledgmentType}{= "",}

\item{ApplicationAcknowledgmentType}{= "",}

\item{CountryCode}{= "",}

\item{CharacterSet}{= "",}

\item{PrincipalLanguageOfMessage}{= "",}

\item{AlternateCharacterSetHandlingScheme}{= "",}

\item{ConformanceStatementID}{= "",}

\item{.FieldSeparator}{= '|',}

\item{.version}{either "2.3.1" or "2.4",}

\item{.trim}{= TRUE}

\item{componentSeparator}{= "^",}

\item{repetitionSeparator}{= "~",}

\item{escapeCharacter}{= "\\",}

\item{subcomponentSeparator}{= "&"}

\item{namespace_ID}{= "",}

\item{universalID}{= "",}

\item{universalIDtype}{= "",}

\item{.sep}{= "^",}

\item{messageType}{= "",}

\item{triggerEvent}{= "",}

\item{messageStructure}{= "",}

\item{versionId}{= '',}

\item{internationalizationCode}{= '',}

\item{internationalversionId}{= '',}
}
\value{
character vector
}
\description{
Convert input to a MSH segment for HL7 2.3.1 or 2.4. Outputs a character vector of
length 1 with input values pipe delimited. Below function definition are helper
functions for particular fields: SendingFacilityComponents, MessageTypeComponents.
}
\section{Functions}{
\itemize{
\item \code{EncodingCharactersComponents}: create a character vector of length 1 with input values separated with a ^

\item \code{SendingFacilityComponents}: create a character vector of length 1 with input values separated with a ^

\item \code{MessageTypeComponents}: create a character vector of length 1 with input values separated with a ^

\item \code{VersionIDComponents}: create a character vector of length 1 with input values separated with a ^
}}

