% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OBR.R
\name{OBR}
\alias{OBR}
\title{Segment OBR: Observation Request}
\usage{
OBR(
  SetID = "",
  PlacerOrderNumber = "",
  FillerOrderNumber = "",
  UniversalServiceID = UniversalServiceIDComponents(),
  PriorityOBR = "",
  RequestedDatetime = "",
  ObservationDateTime = "",
  ObservationEndDateTime = "",
  CollectionVolume = "",
  CollectorIdentifier = "",
  SpecimenActionCode = "",
  DangerCode = "",
  RelevantClinicalInfo = "",
  SpecimenReceivedDateTime = "",
  SpecimenSource = "",
  OrderingProvider = "",
  OrderCallbackPhoneNumber = "",
  PlacerField1 = "",
  PlacerField2 = "",
  FillerField1 = "",
  FillerField2 = "",
  ResultsRptStatusChngDateTime = "",
  ChargetoPractice = "",
  DiagnosticServSetID = "",
  ResultStatus = "",
  ParentResult = "",
  QuantityTiming = "",
  ResultCopiesTo = "",
  Parent = "",
  TransportationMode = "",
  ReasonforStudy = "",
  PrincipalResultInterpreter = "",
  AssistantResultInterpreter = "",
  Technician = "",
  Transcriptionist = "",
  ScheduledDateTime = "",
  NumberofSampleContainers = "",
  TransportLogisticsofCollectedSample = "",
  CollectorsComment = "",
  TransportArrangementResponsibility = "",
  TransportArranged = "",
  EscortRequired = "",
  PlannedPatientTransportComment = "",
  ProcedureCode = "",
  ProcedureCodeModifier = "",
  .FieldSeparator = "|",
  .version = c("2.3.1", "2.4"),
  .trim = TRUE
)
}
\arguments{
\item{SetID}{= "",}

\item{PlacerOrderNumber}{= "",}

\item{FillerOrderNumber}{= "",}

\item{UniversalServiceID}{= UniversalServiceIDComponents(),}

\item{PriorityOBR}{= "",}

\item{RequestedDatetime}{= "",}

\item{ObservationDateTime}{= "",}

\item{ObservationEndDateTime}{= "",}

\item{CollectionVolume}{= "",}

\item{CollectorIdentifier}{= "",}

\item{SpecimenActionCode}{= "",}

\item{DangerCode}{= "",}

\item{RelevantClinicalInfo}{= "",}

\item{SpecimenReceivedDateTime}{= "",}

\item{SpecimenSource}{= "",}

\item{OrderingProvider}{= "",}

\item{OrderCallbackPhoneNumber}{= "",}

\item{PlacerField1}{= "",}

\item{PlacerField2}{= "",}

\item{FillerField1}{= "",}

\item{FillerField2}{= "",}

\item{ResultsRptStatusChngDateTime}{= "",}

\item{ChargetoPractice}{= "",}

\item{DiagnosticServSetID}{= "",}

\item{ResultStatus}{= "",}

\item{ParentResult}{= "",}

\item{QuantityTiming}{= "",}

\item{ResultCopiesTo}{= "",}

\item{Parent}{= "",}

\item{TransportationMode}{= "",}

\item{ReasonforStudy}{= "",}

\item{PrincipalResultInterpreter}{= "",}

\item{AssistantResultInterpreter}{= "",}

\item{Technician}{= "",}

\item{Transcriptionist}{= "",}

\item{ScheduledDateTime}{= "",}

\item{NumberofSampleContainers}{= "",}

\item{TransportLogisticsofCollectedSample}{= "",}

\item{CollectorsComment}{= "",}

\item{TransportArrangementResponsibility}{= "",}

\item{TransportArranged}{= "",}

\item{EscortRequired}{= "",}

\item{PlannedPatientTransportComment}{= "",}

\item{ProcedureCode}{= "",}

\item{ProcedureCodeModifier}{= "",}

\item{.FieldSeparator}{= '|',}

\item{.version}{= c("2.3.1", "2.4"),}

\item{.trim}{= TRUE}
}
\description{
Convert input to a OBR segment for HL7 2.3.1. Outputs a character vector of
length 1 with input values pipe delimited. Below function definition are helper
functions for particular fields: UniversalServiceIDComponents
}
