% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OBX.R
\name{OBX}
\alias{OBX}
\title{Segment OBX: Observation/Result}
\usage{
OBX(
  SetID = "",
  ValueType = "",
  ObservationIdentifier = "",
  ObservationSubID = ObservationIdentifierComponents(),
  ObservationValue = "",
  Units = UnitsComponents(),
  ReferencesRange = "",
  AbnormalFlags = "",
  Probability = "",
  NatureofAbnormalTest = "",
  ObservationResultStatus = "",
  DateLastObsNormalValues = "",
  UserDefinedAccessChecks = "",
  DateTimeoftheObservation = "",
  ProducersID = "",
  ResponsibleObserver = "",
  ObservationMethod = "",
  EquipmentInstanceIdentifier = "",
  DateTimeoftheAnalysis = "",
  .FieldSeparator = "|",
  .version = c("2.3.1", "2.4"),
  .trim = TRUE
)
}
\arguments{
\item{SetID}{= "",}

\item{ValueType}{= "",}

\item{ObservationIdentifier}{= "",}

\item{ObservationSubID}{= ObservationIdentifierComponents(),}

\item{ObservationValue}{= '',}

\item{Units}{= UnitsComponents(),}

\item{ReferencesRange}{= "",}

\item{AbnormalFlags}{= "",}

\item{Probability}{= "",}

\item{NatureofAbnormalTest}{= "",}

\item{ObservationResultStatus}{= "",}

\item{DateLastObsNormalValues}{= "",}

\item{UserDefinedAccessChecks}{= "",}

\item{DateTimeoftheObservation}{= "",}

\item{ProducersID}{= "",}

\item{ResponsibleObserver}{= "",}

\item{ObservationMethod}{= "",}

\item{EquipmentInstanceIdentifier}{= '',}

\item{DateTimeoftheAnalysis}{= '',}

\item{.FieldSeparator}{= '|',}

\item{.version}{= c("2.3.1", "2.4"),}

\item{.trim}{= TRUE}
}
\description{
Convert input to a OBX segment for HL7 2.3.1. Outputs a character vector of
length 1 with input values pipe delimited. Below function definition are helper
functions for particular fields: UnitsComponents
}
