% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORC.R
\name{ORC}
\alias{ORC}
\title{Segment ORC: Common Order}
\usage{
ORC(
  OrderControl = "",
  PlacerOrderNumber = "",
  FillerOrderNumber = "",
  PlacerGroupNumber = "",
  OrderStatus = "",
  ResponseFlag = "",
  QuantityTiming = "",
  Parent = "",
  DateTimeofTransaction = "",
  EnteredBy = "",
  VerifiedBy = "",
  OrderingProvider = "",
  EnterersLocation = "",
  CallBackPhoneNumber = "",
  OrderEffectiveDateTime = "",
  OrderControlCodeReason = "",
  EnteringOrganization = "",
  EnteringDevice = "",
  ActionBy = "",
  AdvancedBeneficiaryNoticeCode = "",
  OrderingFacilityName = "",
  OrderingFacilityAddress = "",
  OrderingFacilityPhoneNumber = "",
  OrderingProviderAddress = "",
  OrderStatusModifier = "",
  .FieldSeparator = "|",
  .version = c("2.3.1", "2.4"),
  .trim = TRUE
)
}
\arguments{
\item{OrderControl}{= '',}

\item{PlacerOrderNumber}{= '',}

\item{FillerOrderNumber}{= '',}

\item{PlacerGroupNumber}{= '',}

\item{OrderStatus}{= '',}

\item{ResponseFlag}{= '',}

\item{QuantityTiming}{= '',}

\item{Parent}{= '',}

\item{DateTimeofTransaction}{= '',}

\item{EnteredBy}{= '',}

\item{VerifiedBy}{= '',}

\item{OrderingProvider}{= '',}

\item{EnterersLocation}{= '',}

\item{CallBackPhoneNumber}{= '',}

\item{OrderEffectiveDateTime}{= '',}

\item{OrderControlCodeReason}{= '',}

\item{EnteringOrganization}{= '',}

\item{EnteringDevice}{= '',}

\item{ActionBy}{= '',}

\item{AdvancedBeneficiaryNoticeCode}{= '',}

\item{OrderingFacilityName}{= '',}

\item{OrderingFacilityAddress}{= '',}

\item{OrderingFacilityPhoneNumber}{= '',}

\item{OrderingProviderAddress}{= '',}

\item{OrderStatusModifier}{= '',}

\item{.FieldSeparator}{= '|',}

\item{.version}{= c("2.3.1", "2.4"),}

\item{.trim}{= TRUE}
}
\description{
Convert input to a segment for HL7 2.3.1. Outputs a character vector of
length 1 with input values pipe delimited. Below function definition are helper
functions for particular fields:
}
