% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_hl7_character_vector.R
\name{parse_hl7_character_vector}
\alias{parse_hl7_character_vector}
\title{Parse a HL7 character vector}
\usage{
parse_hl7_character_vector(hl7_char, name_fields = TRUE, dev_msg = F)
}
\arguments{
\item{hl7_char}{character, a character vector from a valid HL7 file}

\item{name_fields}{boolean, whether to name fields within segements}

\item{dev_msg}{boolean, whether to print package development messages}
}
\value{
a list
}
\description{
The work horse of HL7 parsing
}
