% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_hl7_message.R
\name{parse_hl7_message}
\alias{parse_hl7_message}
\title{Parse a HL7 2.x message}
\usage{
parse_hl7_message(file)
}
\arguments{
\item{file}{A file containing a valid HL7 feed}
}
\value{
a list
}
\description{
Parse a given HL7 file. Function will validate input file and parse both
single and batch HL7 messages.
}
