% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PID.R
\name{PatientAddressComponents}
\alias{PatientAddressComponents}
\title{Components of PID PatientAddress}
\usage{
PatientAddressComponents(
  streetAddress = "",
  otherDesignation = "",
  city = "",
  stateProvince = "",
  zipPostcode = "",
  country = "",
  .sep = "^",
  .trim = TRUE
)
}
\arguments{
\item{streetAddress}{= '',}

\item{otherDesignation}{= '',}

\item{city}{= '',}

\item{stateProvince}{= '',}

\item{zipPostcode}{= '',}

\item{country}{= '',}

\item{.sep}{= "^",}

\item{.trim}{= TRUE}
}
\description{
Function will create a character vector of length 1 with input values separated
with a ^
}
