% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PID.R
\name{PatientIdentifierListComponents}
\alias{PatientIdentifierListComponents}
\title{Components of PID PatientAddress}
\usage{
PatientIdentifierListComponents(
  id = "",
  checkDigit = "",
  codeIdentifyingTheCheckDigitSchemeEmployed = "",
  assigningAuthority = "",
  IdentifierTypeCode = "",
  AssigningFacility = "",
  .sep = "^",
  .trim = TRUE
)
}
\arguments{
\item{id}{= '',}

\item{checkDigit}{= '',}

\item{codeIdentifyingTheCheckDigitSchemeEmployed}{= '',}

\item{assigningAuthority}{= '',}

\item{IdentifierTypeCode}{= '',}

\item{AssigningFacility}{= '',}

\item{.sep}{= "^",}

\item{.trim}{= TRUE}
}
\description{
Function will create a character vector of length 1 with input values separated
with a ^
}
