% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PID.R
\name{PhoneNumberComponents}
\alias{PhoneNumberComponents}
\title{Components of PID PhoneNumber*}
\usage{
PhoneNumberComponents(
  anyText = "",
  telecommunicationUseCode = "",
  telecommunicationEquipmentType = "",
  emailAddress = "",
  countryCode = "",
  areaCityCode = "",
  phoneNumber = "",
  extension = "",
  anyText2 = "",
  .sep = "^",
  .trim = TRUE
)
}
\arguments{
\item{anyText}{= "",}

\item{telecommunicationUseCode}{= "",}

\item{telecommunicationEquipmentType}{= "",}

\item{emailAddress}{= "",}

\item{countryCode}{= "",}

\item{areaCityCode}{= "",}

\item{phoneNumber}{= "",}

\item{extension}{= "",}

\item{anyText2}{= "",}

\item{.sep}{= "^",}

\item{.trim}{= TRUE}
}
\description{
Function will create a character vector of length 1 with input values separated
with a ^
}
