% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.hl7segment}
\alias{print.hl7segment}
\title{Custom print methods for HL7 objects}
\usage{
\method{print}{hl7segment}(x, ...)
}
\arguments{
\item{x}{character, what to print}

\item{..., }{passed onto print

importFrom base print}
}
\description{
Simply to make things print noice.
}
