% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.hl7vector}
\alias{print.hl7vector}
\title{importFrom base print}
\usage{
\method{print}{hl7vector}(x, ...)
}
\arguments{
\item{x}{character, what to print}

\item{..., }{passed onto print}
}
\description{
importFrom base print
}
