% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PV1.R
\name{PV1}
\alias{PV1}
\title{Segment PV1: Patient visit}
\usage{
PV1(
  SetID = "",
  PatientClass = "",
  AssignedPatientLocation = "",
  AdmissionType = "",
  PreadmitNumber = "",
  PriorPatientLocation = "",
  AttendingDoctor = "",
  ReferringDoctor = "",
  ConsultingDoctor = "",
  HospitalService = "",
  TemporaryLocation = "",
  PreadmitTestIndicator = "",
  ReadmissionIndicator = "",
  AdmitSource = "",
  AmbulatoryStatus = "",
  VIPIndicator = "",
  AdmittingDoctor = "",
  PatientType = "",
  VisitNumber = "",
  FinancialClass = "",
  ChargePriceIndicator = "",
  CourtesyCode = "",
  CreditRating = "",
  ContractCode = "",
  ContractEffectiveDate = "",
  ContractAmount = "",
  ContractPeriod = "",
  InterestCode = "",
  TransfertoBadDebtCode = "",
  TransfertoBadDebtDate = "",
  BadDebtAgencyCode = "",
  BadDebtTransferAmount = "",
  BadDebtRecoveryAmount = "",
  DeleteAccountIndicator = "",
  DeleteAccountDate = "",
  DischargeDisposition = "",
  DischargedtoLocation = "",
  DietType = "",
  ServicingFacility = "",
  BedStatus = "",
  AccountStatus = "",
  PendingLocation = "",
  PriorTemporaryLocation = "",
  AdmitDateTime = "",
  DischargeDateTime = "",
  CurrentPatientBalance = "",
  TotalCharges = "",
  TotalAdjustments = "",
  TotalPayments = "",
  AlternateVisitID = "",
  VisitIndicator = "",
  OtherHealthcareProvider = "",
  .FieldSeparator = "|",
  .version = c("2.3.1", "2.4"),
  .trim = TRUE
)
}
\arguments{
\item{SetID}{= "",}

\item{PatientClass}{= "",}

\item{AssignedPatientLocation}{= "",}

\item{AdmissionType}{= "",}

\item{PreadmitNumber}{= "",}

\item{PriorPatientLocation}{= "",}

\item{AttendingDoctor}{= "",}

\item{ReferringDoctor}{= "",}

\item{ConsultingDoctor}{= "",}

\item{HospitalService}{= "",}

\item{TemporaryLocation}{= "",}

\item{PreadmitTestIndicator}{= "",}

\item{ReadmissionIndicator}{= "",}

\item{AdmitSource}{= "",}

\item{AmbulatoryStatus}{= "",}

\item{VIPIndicator}{= "",}

\item{AdmittingDoctor}{= "",}

\item{PatientType}{= "",}

\item{VisitNumber}{= "",}

\item{FinancialClass}{= "",}

\item{ChargePriceIndicator}{= "",}

\item{CourtesyCode}{= "",}

\item{CreditRating}{= "",}

\item{ContractCode}{= "",}

\item{ContractEffectiveDate}{= "",}

\item{ContractAmount}{= "",}

\item{ContractPeriod}{= "",}

\item{InterestCode}{= "",}

\item{TransfertoBadDebtCode}{= "",}

\item{TransfertoBadDebtDate}{= "",}

\item{BadDebtAgencyCode}{= "",}

\item{BadDebtTransferAmount}{= "",}

\item{BadDebtRecoveryAmount}{= "",}

\item{DeleteAccountIndicator}{= "",}

\item{DeleteAccountDate}{= "",}

\item{DischargeDisposition}{= "",}

\item{DischargedtoLocation}{= "",}

\item{DietType}{= "",}

\item{ServicingFacility}{= "",}

\item{BedStatus}{= "",}

\item{AccountStatus}{= "",}

\item{PendingLocation}{= "",}

\item{PriorTemporaryLocation}{= "",}

\item{AdmitDateTime}{= "",}

\item{DischargeDateTime}{= "",}

\item{CurrentPatientBalance}{= "",}

\item{TotalCharges}{= "",}

\item{TotalAdjustments}{= "",}

\item{TotalPayments}{= "",}

\item{AlternateVisitID}{= "",}

\item{VisitIndicator}{= "",}

\item{OtherHealthcareProvider}{= ""}

\item{.FieldSeparator}{= '|'}

\item{.version}{= c("2.3.1", "2.4")}

\item{.trim}{= TRUE}
}
\description{
Convert input to a MSH segment for HL7 2.3.1. Outputs a character vector of
length 1 with input values pipe delimited. Below function definition are helper
functions for particular fields: SendingFacilityComponents, MessageTypeComponents.
}
