% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_names.R
\docType{data}
\name{coded_element}
\alias{coded_element}
\alias{extended_telecommunication_number}
\alias{segment_names}
\title{Segment field names}
\format{
An object of class \code{character} of length 6.

An object of class \code{character} of length 9.

An object of class \code{list} of length 23.
}
\usage{
coded_element

extended_telecommunication_number

segment_names
}
\description{
A list of segment (MSH, PID, etc) and sub-field names (MessageType,
PatientIdentifierList, etc). Not exhaustive. Used with \code{add_field_names()},
which simply matches by \code{names}. Additional list objects are used for
repetitive sub-field names, so far: \code{coded_element}, \code{extended_telecommunication_number},
and new ones included whenever the need arises.
Coded Element (CE) field structure
}
\section{Functions}{
\itemize{
\item \code{coded_element}: A list of field names for coded element (CE) fields. Used by \code{segment_names}
This is a recurring structure used to identify a value in one or possibly
two different coding systems, e.g. a lab might transmit the same value in both
their local coding and also the universal LOINC standard code.
Described further at: http://www.hl7.eu/HL7v2x/v231/hl7v231typCE.htm

\item \code{extended_telecommunication_number}: A list of field names for Extended Telecommunication Number (XTN) fields. Used by \code{segment_names}

\item \code{segment_names}: The actual list of segment (MSH, PID, etc) and sub-field names (MessageType,
PatientIdentifierList, etc), that calls upon other lists described here.
}}

\keyword{datasets}
