% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_segment.R
\name{split_segment}
\alias{split_segment}
\title{Split a HL7 segment into respective fields}
\usage{
split_segment(
  segment,
  FieldSeparator = "|",
  ComponentSeparator = "^",
  SubcomponentSeparator = "&",
  FieldRepeatSeparator = "~"
)
}
\arguments{
\item{segment}{a character vector}

\item{FieldSeparator}{character, splits this}

\item{ComponentSeparator}{character, splits this}

\item{SubcomponentSeparator}{character, splits this}

\item{FieldRepeatSeparator}{character, splits this}
}
\description{
This function simply does a \code{strsplit()} routine, and will do this for
sub-components as well.
}
