% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit_as_list.R
\name{strsplit_as_list}
\alias{strsplit_as_list}
\title{Helper function for strsplit}
\usage{
strsplit_as_list(x, split)
}
\arguments{
\item{x}{a character vector}

\item{split}{character, what to split, passed on to strsplit}
}
\description{
Helps extract the result of \code{strsplit} and then return it as a list
}
