% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_trailing_chars.R
\name{trim_trailing_chars}
\alias{trim_trailing_chars}
\title{Trim trailing characters from character vector}
\usage{
trim_trailing_chars(x, char)
}
\arguments{
\item{x}{character, string to trim}

\item{char}{character, character to trim from end of string}
}
\description{
Trim trailing characters from character vector
}
