% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OBX.R
\name{UnitsComponents}
\alias{UnitsComponents}
\title{Components of OBR Units}
\usage{
UnitsComponents(
  identifier = "",
  text = "",
  nameOfCodingSystem = "",
  alternateIdentifier = "",
  alternateText = "",
  nameOfAlternateCodingSystem = "",
  .sep = "^",
  .trim = TRUE
)
}
\arguments{
\item{identifier}{= '',}

\item{text}{= '',}

\item{nameOfCodingSystem}{= '',}

\item{alternateIdentifier}{= '',}

\item{alternateText}{= '',}

\item{nameOfAlternateCodingSystem}{= '',}

\item{.sep}{= '^',}

\item{.trim}{= TRUE}
}
\description{
Function will create a character vector of length 1 with input values separated
with a ^
}
