% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_hl7.R
\name{validate_hl7}
\alias{validate_hl7}
\title{Check the headers of a HL7 file}
\usage{
validate_hl7(file)
}
\arguments{
\item{file}{input file}
}
\description{
Function looks for HL7 headers - FHS, BHS and MSH - in a file and responds
appropriately.
}
