% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OBX.R
\name{variedComponents}
\alias{variedComponents}
\title{Components of ObservationIdentifier}
\usage{
variedComponents(..., .sep = "^")
}
\arguments{
\item{...}{dots!}

\item{.sep}{= '^'}
}
\description{
Function will create a character vector of length 1 with input values separated
with a ^
}
